+++ apps/autotest/models/TST/public/locale/lang-en.js
@@ -1,9 +1,0 @@
-UB.i18nExtend({
-  general: 'General',
-  performing: 'Performing',
-  myIsuuesLocale: {
-    one: 'One',
-    two: 'Two'
-  },
-  dashboard: 'Dashboard'
-})